Created VideosContext so that video data can be accessed throughout the app without having to pass in props.
