+++ pixelflix-app/src/App.jsx
@@ -4,6 +4,7 @@ import Signup from "./pages/Signup/Signup";
 import Login from "./pages/Login/Login";
 import Home from "./pages/Home/Home";
 import { useAuth } from "./context/AuthContext";
+import { VideoProvider } from "./context/VideosContext";
 
 function App() {
   const { currentUser } = useAuth();
@@ -13,18 +14,20 @@ function App() {
   };
 
   return (
-    <Routes>
-      <Route path="/signup" element={<Signup />} />
-      <Route path="/login" element={<Login />} />
-      <Route
-        path="/*"
-        element={
-          <RequireAuth>
-            <Home />
-          </RequireAuth>
-        }
-      />
-    </Routes>
+    <VideoProvider>
+      <Routes>
+        <Route path="/signup" element={<Signup />} />
+        <Route path="/login" element={<Login />} />
+        <Route
+          path="/*"
+          element={
+            <RequireAuth>
+              <Home />
+            </RequireAuth>
+          }
+        />
+      </Routes>
+    </VideoProvider>
   );
 }
