+++ pixelflix-app/src/pages/Login/Login.jsx
@@ -49,19 +49,6 @@ function Login() {
       return;
     }
 
-    // Check if the email is already registered
-    try {
-      const signInMethods = await fetchSignInMethodsForEmail(auth, email);
-      // if (signInMethods.length === 0) {
-      //   setEmailNonexistent(true);
-      //   console.log("No account found with this email address.");
-      //   return;
-      // }
-    } catch (err) {
-      console.error(err);
-      return;
-    }
-
     if (password === "") {
       setPasswordEmpty(true);
       return;
